shall_we_meet/requirements.txt
asgiref==3.2.3
astroid==2.3.3
backcall==0.1.0
decorator==4.4.1
Django==3.0.1
django-crispy-forms==1.8.1
django-tempus-dominus==5.1.2.13
entrypoints==0.3
flake8==3.7.9
isort==4.3.21
jedi==0.15.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.2
psycopg2-binary==2.8.4
ptyprocess==0.6.0
public==2019.4.13
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.5.2
pylint==2.4.4
pytz==2019.3
self==2019.4.13
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
typed-ast==1.4.1
wcwidth==0.1.7
wrapt==1.11.2
